% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TJU_Workday.R
\name{allDates}
\alias{allDates}
\alias{allDates.integer}
\alias{allDates.character}
\alias{allDates.yearmon}
\alias{allDates.yearqtr}
\title{All \link[base]{Date}s in a Time Interval}
\usage{
allDates(x)

\method{allDates}{integer}(x)

\method{allDates}{character}(x)

\method{allDates}{yearmon}(x)

\method{allDates}{yearqtr}(x)
}
\arguments{
\item{x}{R objects, such as
\describe{
\item{\link[base]{integer}}{
year, e.g., \code{x = 2020L} returns all \link[base]{Date}s from 2020-01-01 to 2020-12-31}
\item{\link[zoo]{yearmon}}{
year-month object from package \CRANpkg{zoo}}
\item{\link[zoo]{yearqtr}}{
year-quarter object from package \CRANpkg{zoo}}
\item{\link[base]{character}}{
convertible to \link[zoo]{yearmon} or \link[zoo]{yearqtr} object}
}}
}
\value{
Function \link{allDates} returns a \link[base]{Date} \link[base]{vector}.
}
\description{
Find all \link[base]{Date}s in a time interval.
}
\details{
Function \link{allDates} returns all \link[base]{Date}s in a given time interval.
}
\keyword{internal}
