% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibentry2rmd.R
\name{bibentry2rmd}
\alias{bibentry2rmd}
\title{R Markdown Format of \link[utils]{citation} and/or \link[utils]{bibentry}}
\usage{
bibentry2rmd(x = "R")
}
\arguments{
\item{x}{\link[base]{character} scalar,
\code{'R'} (default) or name of an R package}
}
\value{
Function \link{bibentry2rmd} returns a \link[base]{character} scalar or \link[base]{vector}.
}
\description{
R markdown format of a \link[utils]{citation} and/or \link[utils]{bibentry} object.
}
\details{
Function \link{bibentry2rmd} beautifies the output from
function \code{utils:::format.bibentry} (with option \code{style = 'text'})
in the following ways.
\itemize{
\item{Line break \code{'\\n'} is replaced by a white space;}
\item{Fancy quotes \eqn{``}, \eqn{''}, \eqn{`} and \eqn{'} are removed;}
\item{doi entries are shown as URLs with labels (in R markdown grammar).}
}
}
\examples{
bibentry2rmd('survival')
if (FALSE) { # disabled for ?devtools::check
ap = rownames(installed.packages())
lapply(ap, FUN = bibentry2rmd)
}
}
