% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageKey.R
\name{bibentry_key<-}
\alias{bibentry_key<-}
\title{Assign Key to \link[utils]{bibentry}}
\usage{
bibentry_key(x) <- value
}
\arguments{
\item{x}{\link[utils]{bibentry} (and inherited \link[utils]{citation}) object}

\item{value}{\link[base]{character} scalar or \link[base]{vector}}
}
\description{
Assign an \link[base]{attributes} named \code{'key'} to \link[utils]{bibentry} object.
}
\note{
\code{'key'} is the attribute of \link[utils]{bibentry}, which is used in
\code{utils:::.bibentry_get_key} and \code{utils:::toBibtex.bibentry}.
}
