% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_confint.R
\name{bin_confint_}
\alias{bin_confint_}
\title{Clopper-Pearson Exact Binomial Confidence Interval}
\usage{
bin_confint_(
  x,
  n,
  level = 0.95,
  alternative = c("two.sided", "less", "greater"),
  ...
)
}
\arguments{
\item{x}{positive \link[base]{integer} \link[base]{vector}, counts}

\item{n}{positive \link[base]{integer} \link[base]{vector}, sample sizes \eqn{n}}

\item{level}{\link[base]{numeric} scalar, confidence level, default .95}

\item{alternative}{\link[base]{character} scalar,
\code{'two.sided'} (default), \code{'less'} or \code{'greater'}}

\item{...}{potential parameters}
}
\value{
Function \link{bin_confint_} returns a \link[base]{matrix}.
}
\description{
Clopper-Pearson exact binomial confidence interval.
}
\note{
Function \code{Hmisc::binconf} uses \link[stats]{qf}.

\link[stats]{binom.test} and \code{binom::binom.confint} uses \link[stats]{qbeta} (equivalent but much cleaner!)

Only \link[stats]{binom.test} provides one-sided confidence interval.
}
\examples{
bin_confint_(0:10, 10L)
bin_confint_(0:10, 10L, alternative = 'less')
bin_confint_(0:10, 10L, alternative = 'greater')

}
\references{
\doi{10.1093/biomet/26.4.404}
}
