% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom_confint.R
\name{binom_confint}
\alias{binom_confint}
\title{Clopper-Pearson Exact Binomial Confidence Interval}
\usage{
binom_confint(
  x,
  n,
  conf.level = 0.95,
  alternative = c("two.sided", "less", "greater"),
  ...
)
}
\arguments{
\item{x}{positive \link[base]{integer} scalar or \link[base]{vector}, counts}

\item{n}{positive \link[base]{integer} scalar or \link[base]{vector}, sample sizes \eqn{n}}

\item{conf.level, alternative, ...}{additional parameters of function \link[stats]{binom.test}}
}
\value{
Function \code{\link[=binom_confint]{binom_confint()}} returns a \link[base]{data.frame}.
}
\description{
Clopper-Pearson exact binomial confidence interval.
}
\examples{
binom_confint(0:10, 10L)
binom_confint(0:10, 10L, alternative = 'less')
binom_confint(0:10, 10L, alternative = 'greater')
}
\keyword{internal}
