% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDuplicated.R
\name{checkDuplicated}
\alias{checkDuplicated}
\title{checkDuplicated}
\usage{
checkDuplicated(data, f, file_duplicated = tempfile(fileext = ".txt"), ...)
}
\arguments{
\item{data}{\link[base]{data.frame}}

\item{f}{\link[stats]{formula}}

\item{file_duplicated}{\link[base]{character} scalar}

\item{...}{additional parameters, currently not in use}
}
\value{
Function \link{checkDuplicated} returns a \link[base]{data.frame}.
}
\description{
..
}
\examples{
(d1 = data.frame(A = c(1, 1), B = c(NA_character_, 'text')))


(d2 = data.frame(A = c(1, 2), B = c(NA_character_, 'text')))

}
