% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.R
\name{force_bool}
\alias{force_bool}
\title{Force to \link[base]{logical} \link[base]{vector}}
\usage{
force_bool(x, else_return = x)
}
\arguments{
\item{x}{a \link[base]{factor}, or a \link[base]{vector}}

\item{else_return}{an R object to return if cannot force into \link[base]{logical}.
Default \code{x}.}
}
\description{
..
}
\details{
Function \link{force_bool} tries to turn an object into \link[base]{logical}.
}
\examples{
force_bool(c('0', '1', '0', NA))
(tmp = factor(rep(0:1, times = 10L)))
force_bool(tmp)

}
