% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exact_confint.R
\name{format.exact_confint}
\alias{format.exact_confint}
\title{Format Clopper-Pearson Exact Binomial Confidence Interval}
\usage{
\method{format}{exact_confint}(x, data.name, ...)
}
\arguments{
\item{x}{an \link{exact_confint}}

\item{data.name}{(optional) \link[base]{character} scalar}

\item{...}{additional parameters, currently not in use}
}
\value{
Function \link{format.exact_confint} returns a \link[base]{character}
\link[base]{vector} when argument \code{data.name} is \link[base]{missing};
otherwise, a \link[base]{noquote} \link[base]{character} \link[base]{matrix} is returned.
}
\description{
To \link[base]{format} a Clopper-Pearson exact binomial confidence interval.
}
