% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_confint.R
\name{format_bin_confint_}
\alias{format_bin_confint_}
\title{Format Clopper-Pearson Exact Binomial Confidence Interval}
\usage{
format_bin_confint_(..., count.name)
}
\arguments{
\item{...}{see function \link{bin_confint_}}

\item{count.name}{(optional) \link[base]{character} scalar}
}
\value{
Function \link{format_bin_confint_} returns a \link[base]{character}
\link[base]{vector} when argument \code{count.name} is \link[base]{missing};
otherwise, a \link[base]{noquote} \link[base]{character} \link[base]{matrix} is returned.
}
\description{
..
}
