% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_file_size.R
\name{format_file_size}
\alias{format_file_size}
\title{Format \link[base]{file.size}}
\usage{
format_file_size(..., units = "auto")
}
\arguments{
\item{...}{\link[base]{character} scalar or \link[base]{vector}, file paths}

\item{units}{\link[base]{character} scalar,
see parameter \code{units} of function \link[utils]{format.object_size}}
}
\value{
Function \link{format_file_size} returns a \link[base]{character} \link[base]{vector}.
}
\description{
Format \link[base]{file.size}.
}
\details{
Function \link{format_file_size} formats \link[base]{file.size}
in the same manner as
function \link[utils]{format.object_size} does to \link[utils]{object.size}.
}
\note{
Return of function \link[base]{file.size} is simply \link[base]{numeric},
thus we will not be able to define an S3 method dispatch for the S3 generic \link[base]{format}, yet.
}
\examples{
# format_file_size('./R/allequal.R', './R/approxdens.R')
}
