% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_named.R
\name{format_named}
\alias{format_named}
\title{format_named}
\usage{
format_named(x, sep = ": ")
}
\arguments{
\item{x}{\link[base]{character} \link[base]{vector},
or a \link[base]{list} of \link[base]{character} object.
Input \code{x} must be named}

\item{sep}{\link[base]{character} scalar, see \link[base]{paste}}
}
\value{
Function \code{\link[=format_named]{format_named()}} returns a \link[base]{character} \link[base]{vector}.
}
\description{
format_named
}
\examples{
c(a = 'a1', bc = '2\n3') |> format_named()

list(a = '1\n2', b = character(), cd = '3\n4', efg = '5\n6\n7') |> format_named()

c(a = '1\n2') |> format_named()

}
\keyword{internal}
