% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex_ext.R
\name{grepl_}
\alias{grepl_}
\title{\link[base]{grepl}, but recognize \code{NA}}
\usage{
grepl_(x, ...)
}
\arguments{
\item{x}{\link[base]{character} \link[base]{vector}}

\item{...}{additional parameters of function \link[base]{grepl}}
}
\description{
\link[base]{grepl}, but recognize \code{NA}
}
\examples{
grepl_(pattern = '^a', x = c('a', 'b', NA_character_))
}
