% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex_ext.R
\name{gsub_yes}
\alias{gsub_yes}
\title{gsub_yes}
\usage{
gsub_yes(pattern, replacement = "YES", x, ...)
}
\arguments{
\item{pattern}{\link[base]{character} scalar}

\item{replacement}{\link[base]{character} scalar}

\item{x}{\link[base]{character} \link[base]{vector}}

\item{...}{additional parameters of function \link[base]{grepl}}
}
\value{
Function
}
\description{
gsub_yes
}
\examples{
gsub_yes('^a', x = c('a', 'b', NA_character_))
}
