% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_.R
\name{inspect_}
\alias{inspect_}
\title{Data Inspection}
\usage{
inspect_(x, duplicated_rm = TRUE, date_pattern = "^Date_", ...)
}
\arguments{
\item{x}{\link[base]{data.frame}}

\item{duplicated_rm}{\link[base]{logical} scalar, whether to remove duplicated \emph{rows}.
Default \code{FALSE}.}

\item{date_pattern}{regular expression \link[base]{regex},
pattern of column names for \link[base]{Dates}.
Default \verb{^Date_}.}

\item{...}{additional parameters, currently not in use}
}
\value{
Function \link{inspect_} returns a \link[base]{data.frame}.
}
\description{
..
}
\details{
\itemize{
\item {convert all eligible \link[base]{integer}, \link[base]{numeric},
\link[base]{factor} and \link[base]{character} columns to \link[base]{logical}.}
}
}
\note{
Be aware of potential name clash, e.g., \code{lavaan::inspect}.
}
