% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction.formula.R
\name{interaction.formula}
\alias{interaction.formula}
\title{A \link[stats]{formula} Interface for \link[base]{interaction}}
\usage{
interaction.formula(formula, data, ...)
}
\arguments{
\item{formula}{\link[stats]{formula}}

\item{data}{\link[base]{data.frame}}

\item{...}{parameters \code{drop}, \code{sep} and \code{lex.order} of function \link[base]{interaction}}
}
\description{
A \link[stats]{formula} interface for function \link[base]{interaction}.
}
\details{
Function \link{interaction.formula} is useful when the input \link[base]{factor}s are
specified as a \link[stats]{formula} based on a \link[base]{data.frame}, e.g., in function \link[base]{split.data.frame}.
}
\note{
R base function \link[base]{interaction} correctly handles syntactically invalid names (see \link[base]{make.names}).
}
\examples{
data(farms, package = 'MASS')
stopifnot(identical(with(farms, interaction(Mois, Manag)), 
  interaction.formula(~ Mois + Manag, data = farms)))
stopifnot(identical(with(farms, interaction(Mois, Manag, drop = TRUE)),
  interaction.formula(~ Mois + Manag, data = farms, drop = TRUE)))
  
# syntactically invalid names, okay!
(x = data.frame('a b' = 1:3, 'c d' = LETTERS[1:3], check.names = FALSE))
(ret = interaction.formula(~ `a b` + `c d`, data = x, drop = TRUE))
stopifnot(identical(ret, with(x, interaction(`a b`, `c d`, drop = TRUE))))
}
