% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lang_tabs.R
\name{lang_tabs}
\alias{lang_tabs}
\title{Cross-Tabulation by a \link[base]{list} of \link[base]{language}s}
\usage{
lang_tabs(lang, data, ...)
}
\arguments{
\item{lang}{a \link[base]{list} of \link[base]{language} objects}

\item{data}{a \link[base]{data.frame}}

\item{...}{additional parameters of function \link[base]{table}}
}
\description{
Cross-Tabulation by a \link[base]{list} of \link[base]{language}s
}
\seealso{
\link[stats]{xtabs}
}
\keyword{internal}
