% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ma.R
\name{ma}
\alias{ma}
\title{Moving Average}
\usage{
ma(x, order = 5L)
}
\arguments{
\item{x}{\link[base]{numeric} \link[base]{vector}}

\item{order}{\link[base]{integer} scalar}
}
\value{
Function \code{\link[=ma]{ma()}} returns a time-series \link[stats]{ts} object from workhorse function \link[stats]{filter}.
}
\description{
..
}
\note{
Function \code{\link[=ma]{ma()}} is a simplified version of function \link[forecast]{ma} in package \CRANpkg{forecast}.

Function \code{\link[=ma]{ma()}} is much faster than function \code{zoo::rollmean()}.

Function \code{\link[=ma]{ma()}} imports function \code{stats::filter()},
\emph{\strong{not}} function \code{dplyr::filter()}.
}
\examples{
ma(1:20, order = 3L) |> unclass()
ma(1:2, order = 3L) |> unclass()
}
\references{
\url{https://stackoverflow.com/questions/743812/calculating-moving-average}
}
