% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmelt.R
\name{mmelt}
\alias{mmelt}
\title{\link[reshape2]{melt} by Multiple Groups of Measurement Variables}
\usage{
mmelt(data, id.vars, measure_rx, variable.name = "variable")
}
\arguments{
\item{data}{a \link[base]{data.frame}}

\item{id.vars}{\link[base]{character} \link[base]{vector}, see function \link[reshape2]{melt.data.frame}.
Default is all variables not matched by \code{measure_rx}.}

\item{measure_rx}{\emph{named} \link[base]{character} \link[base]{vector},
\link[base]{regex}s to determine \code{measure.vars} of function \link[reshape2]{melt.data.frame},
while \code{names(measure_rx)} serve
as \code{value.name} of function \link[reshape2]{melt.data.frame}}

\item{variable.name}{\link[base]{character} scalar,
see function \link[reshape2]{melt.data.frame}}
}
\value{
Function \link{mmelt} returns a \link[base]{data.frame}
}
\description{
\link[reshape2]{melt} by multiple groups of measurement variables.
}
\details{
Function \link{mmelt} melts by multiple groups of measurement variables.
}
\examples{
(iris0 = iris[c(1:3, 51:53, 101:103),])
mmelt(iris0, measure_rx = c(len = '\\\\.Length$', wid = '\\\\.Width$'), variable.name = 'Part')
}
