% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_adjust_.R
\name{p_adjust_}
\alias{p_adjust_}
\title{Perform All \eqn{p}-value Adjustments}
\usage{
p_adjust_(p)
}
\arguments{
\item{p}{\link[base]{numeric} \link[base]{vector}}
}
\value{
Function \link{p_adjust_} returns a \link[base]{matrix} of adjusted p-values using all available \link[stats]{p.adjust.methods}.
}
\description{
Perform all \eqn{p}-value adjustments.
}
\details{
Function \link{p_adjust_} speeds up function \link[stats]{p.adjust}.
}
\examples{
\donttest{for (i in seq_len(10L)) {
 n = sample(20:30, size = 1L)
 p = runif(n, min = .Machine$double.eps, max = 1 - .Machine$double.eps)
 (tmp = p_adjust_(p))
 stopifnot(identical(tmp, cbind(
  holm = p.adjust(p, 'holm'),
  hochberg = p.adjust(p, 'hochberg'),
  hommel = p.adjust(p, 'hommel'),
  bonferroni = p.adjust(p, 'bonferroni'),
  BH = p.adjust(p, 'BH'),
  BY = p.adjust(p, 'BY'),
  none = p)))
}}
}
