% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageKey.R
\name{packageKey}
\alias{packageKey}
\title{Use \link[utils]{packageName} as \link[utils]{citation} Key}
\usage{
packageKey(x, overwrite = FALSE)
}
\arguments{
\item{x}{\link[utils]{citation} object}

\item{overwrite}{\link[base]{logical} scalar, whether to overwrite
default key(s).  Default \code{FALSE}.}
}
\value{
Function \link{packageKey} returns a \link[utils]{citation} object.
}
\description{
Use \link[utils]{packageName} as \link[utils]{citation} key
}
\details{
Function \link{packageKey} adds \link[utils]{packageName} as \link[utils]{citation} key.
}
\note{
As of June 2024:
if the last call to \link[utils]{bibentry} in
function \link[utils]{citation} adds the argument \code{key = package},
or provide a parameter to allow end-user to enable such choice,
then we don't need the function \link{packageKey}.
}
\examples{
if (FALSE) {
 ap = installed.packages()
 table(aa[, 'Priority'])
 which(ap[, 'Priority'] == 'recommended')
}

toBibtex(ct <- citation('survival')) # has default key(s)
toBibtex(packageKey(ct))
toBibtex(packageKey(ct)[1L])
toBibtex(packageKey(ct, overwrite = TRUE)[2L])

toBibtex(ct <- citation()) # no default key(s)
toBibtex(packageKey(ct))
}
