% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phone10_zip5.R
\name{phone10}
\alias{phone10}
\title{10-digit US phone number}
\usage{
phone10(x, sep = "")
}
\arguments{
\item{x}{\link[base]{character} \link[base]{vector}}

\item{sep}{\link[base]{character} scalar}
}
\value{
Function \link{phone10} returns a \link[base]{character} \link[base]{vector} of \link[base]{nchar}-10.
}
\description{
..
}
\details{
Function \link{phone10} converts all US and Canada (+1) phone numbers to 10-digit.
}
\examples{
x = c('+16108885188', '15108581787', '8588889426')
phone10(x)
phone10(x, sep = '-')

}
