% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.R
\name{POSIXct2difftime}
\alias{POSIXct2difftime}
\title{Convert No-Date \link[base]{POSIXct} to \link[base]{difftime}}
\usage{
POSIXct2difftime(
  x,
  else_return = stop("Expecting all '1899-12-31 hr:min:sec UTC'")
)
}
\arguments{
\item{x}{\link[base]{POSIXct}}

\item{else_return}{exception handling}
}
\description{
..
}
\details{
\CRANpkg{readxl} will read \code{'hour:min:sec'} as \code{'1899-12-31 hr:min:sec UTC'}
}
\seealso{
\code{lubridate:::year.default} \code{lubridate:::tz.POSIXt}
}
