% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_exception.R
\name{prop_exception}
\alias{prop_exception}
\alias{prop_missing}
\alias{prop_nonmissing}
\title{Proportions of Exceptions}
\usage{
prop_missing(x, data.name = deparse1(substitute(x)), ...)

prop_nonmissing(x, data.name = deparse1(substitute(x)), ...)
}
\arguments{
\item{x}{..}

\item{data.name}{..}

\item{...}{all potential parameters of \link[base]{as.data.frame}}
}
\description{
..
}
\details{
..
}
\examples{
prop_missing(swiss)
prop_missing(airquality)
prop_missing(airquality$Ozone)

}
