% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reduce_call.R
\name{Reduce_call}
\alias{Reduce_call}
\alias{Reduce_plus}
\alias{Reduce_average}
\title{Create a \link[base]{call} of Plus}
\usage{
Reduce_plus(x)

Reduce_average(x)
}
\arguments{
\item{x}{\link[base]{character} \link[base]{vector}}
}
\value{
Function \link{Reduce_plus} returns a \link[base]{call}
}
\description{
Create a \link[base]{call} of Plus
}
\examples{
Reduce_plus(letters[1:3])

Reduce_average(letters[1:3])
}
\keyword{internal}
