% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_rx.R
\name{select_rx}
\alias{select_rx}
\title{Select by Regular Expression}
\usage{
select_rx(pattern, envir = parent.frame(), ...)
}
\arguments{
\item{pattern}{\link[base]{character} scalar, regular expression}

\item{envir}{an \link[base]{environment},
a \link[base]{list},
a \link[base]{data.frame},
or a \link[base]{matrix}.}

\item{...}{additional parameters of \link[base]{grep}, most importantly \code{invert}}
}
\value{
Function \link{select_rx} selects

\itemize{
\item {columns from a \link[base]{data.frame} by \code{pattern} matching against its \link[base]{names},
and returns a \link[base]{data.frame}.}
\item {elements from a \link[base]{list} by \code{pattern} matching against its \link[base]{names},
and returns a \link[base]{list}.}
\item {columns from a \link[base]{matrix} by \code{pattern} matching against its \link[base]{colnames},
and returns a \link[base]{matrix}.}
\item {objects in an \link[base]{environment} by \code{pattern} matching against the names of variables within,
and returns a \strong{\link[base]{list}}
instead of an \link[base]{environment}.}
}
}
\description{
..
}
\examples{
head(select_rx('\\\\.Width$', envir = iris))
head(select_rx('\\\\.Length$', envir = iris, invert = TRUE))

select_rx('^Rural', VADeaths)

with(head(iris), expr = {
 select_rx(pattern = '\\\\.Width$')
})
}
