% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_xor.R
\name{set_xor}
\alias{set_xor}
\title{Exclusive-OR Elements}
\usage{
set_xor(
  e1,
  e2,
  name1 = deparse1(substitute(e1)),
  name2 = deparse1(substitute(e2))
)
}
\arguments{
\item{e1, e2}{two R objects of the same \link[base]{typeof}}

\item{name1, name2}{(optional) \link[base]{character} scalars,
human-friendly names of \code{e1} and \code{e2}.
Default is the function call of \code{e1} and \code{e2}.}
}
\value{
Function \code{\link[=set_xor]{set_xor()}} returns either a \link[base]{list} or a \link[base]{vector}.
}
\description{
Exclusive-OR elements in two vectors.
}
\details{
Function \code{\link[=set_xor]{set_xor()}} returns the exclusive-OR elements in each of the sets, which is slow
and only intended for end-user.
}
\examples{
set_xor(1:5, 3:7)
set_xor(1:5, 1:3)
}
\seealso{
\link[base]{setequal}
}
\keyword{internal}
