% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourcePath.R
\name{sourcePath}
\alias{sourcePath}
\title{Source All R Files under a Directory}
\usage{
sourcePath(path, ...)
}
\arguments{
\item{path}{\link[base]{character} scalar, parent directory of \code{.R} files}

\item{...}{additional parameters of \link[base]{source}}
}
\value{
Function \link{sourcePath} does not have a returned value
}
\description{
\link[base]{source} all \verb{*.R} and \verb{*.r} files under a directory.
}
