% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitKey.R
\name{split_ext}
\alias{split_ext}
\alias{splitKey}
\alias{splitOrd}
\title{Split \link[base]{character} \link[base]{vector}}
\usage{
splitKey(x, key = xkey, data.name = substitute(x), envir = parent.frame(), ...)

splitOrd(
  x,
  nm = stop("must specify new names"),
  data.name = substitute(x),
  envir = parent.frame(),
  ...
)
}
\arguments{
\item{x}{\link[base]{character} \link[base]{vector}, each element being a set of
keywords separated by a symbol (e.g., \code{','})}

\item{key}{(optional for function \link{splitKey}) \link[base]{character} \link[base]{vector},
user-specified keywords.  Default to all keywords appearing in input \code{x}}

\item{data.name}{(optional) \link[base]{character} scalar or \link[base]{name},
name of input \code{x}}

\item{envir}{\code{NULL} or \link[base]{environment}.
If an \link[base]{environment} is specified, then assign the result to it
(i.e., when used inside \link[base]{within.data.frame}).}

\item{...}{potential parameters of \link[base]{strsplit}, most importantly \code{split}}

\item{nm}{(for function \link{splitOrd}) \link[base]{character} \link[base]{vector}}
}
\value{
Function \link{splitKey} returns a \link[base]{logical} \link[base]{matrix} if \code{assign2parent = FALSE}.
Otherwise the \link[base]{logical} \link[base]{vector}s are assigned to the parent frame
(i.e., when used inside \link[base]{within.data.frame}).

Function \link{splitOrd} returns a \link[base]{logical} \link[base]{matrix} if \code{envir = NULL}.
Otherwise the \link[base]{logical} \link[base]{vector}s are assigned to the \code{envir}
(i.e., when used inside \link[base]{within.data.frame}).
}
\description{
Split \link[base]{character} \link[base]{vector}, into keywords or by order of appearance.
}
\details{
Function \link{splitKey} finds out whether each keyword appears in each element of input \code{x}.

\code{NA_character_} or \code{''} entries in input \code{x} are regarded
as negative (i.e., none of the keywords exists),
instead of as missingness (i.e., we do not know if any of the keywords exists).
This practice is most intuitive to clinicians.
}
\examples{

letters[1:4] |> splitKey(split = ';;', envir = NULL) # exception

x = c('a,b,', 'c,a,b,,', NA_character_, '', 'a,b,a')
x |> splitKey(split = ',', envir = NULL)

data.frame(x) |> within.data.frame(expr = splitKey(x, split = ','))

data.frame(x) |> within.data.frame(expr = splitKey(x, split = ',', data.name = 'cancer'))

if (FALSE) {
library(microbenchmark)
X = rep(x, times = 10L)
microbenchmark( # speed O(n)
 splitKey(x, split = ',', envir = NULL), 
 splitKey(X, split = ',', envir = NULL))
}
x = c('T2,N0,M0,B1', 'T4,N0, ,B0', 'T2,N0,M0,B0', 'T2,N1,M0,B0', ',Nx,M0,B0',
  'T4,N3,M0,B2', NA, 'T4,N1,M0,B2')
nm = c('T', 'N', 'M', 'B')
splitOrd(x, split = ',', nm = nm, envir = NULL)
  
data.frame(x) |> within.data.frame(expr = {
  splitOrd(x, split = ',', nm = nm, data.name = 'Stage')
})
  
}
