% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitKey.R
\name{splitKey}
\alias{splitKey}
\title{Split \link[base]{character} \link[base]{vector} into Keywords}
\usage{
splitKey(x, keys = xkeys, data.name = substitute(x), assign2parent = TRUE, ...)
}
\arguments{
\item{x}{\link[base]{character} \link[base]{vector}, each element being a collection of
keywords separated by a symbol (e.g., \code{','})}

\item{keys}{(optional) \link[base]{character} \link[base]{vector}, user-specified
keywords.  Default to all keywords appearing in input \code{x}}

\item{data.name}{(optional) \link[base]{character} scalar or \link[base]{name},
name of input \code{x}}

\item{assign2parent}{\link[base]{logical} scalar, whether to assign the result to the parent frame
(i.e., when used inside \link[base]{within.data.frame}).  Default \code{TRUE}}

\item{...}{potential parameters of \link[base]{strsplit}, most importantly \code{split}}
}
\value{
Function \link{splitKey} returns a \link[base]{logical} \link[base]{matrix} if \code{assign2parent = FALSE}.
Otherwise the \link[base]{logical} \link[base]{vector}s are assigned to the parent frame
(i.e., when used inside \link[base]{within.data.frame}).
}
\description{
..
}
\details{
Function \link{splitKey} finds out whether each keyword appears in each element of input \code{x}.

\code{NA_character_} or \code{''} entries in input \code{x} are regarded
as negative (i.e., none of the key words exists),
instead of as missingness (i.e., we do not know if any of the keywords exists).
This practice is most intuitive to clinicians.
}
\note{
It is presumed there is few duplication in the input \code{x}.
}
\examples{

x = c('a,b,', 'c,a,b,,', NA_character_, '', 'a,b,')
splitKey(x, split = ',', assign2parent = FALSE)

within(data.frame(x), expr = {
  splitKey(x, split = ',')
  splitKey(x, split = ',', data.name = 'cancer')
})

if (FALSE) {
library(microbenchmark)
X = rep(x, times = 1e1L)
microbenchmark( # speed O(n)
 splitKey(x, split = ',', assign2parent = FALSE), 
 splitKey(X, split = ',', assign2parent = FALSE))
}


}
