% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_.R
\name{style_basename}
\alias{style_basename}
\title{Highlight Style for File Base Name}
\usage{
style_basename(x)
}
\arguments{
\item{x}{\link[base]{character} scalar}
}
\value{
Function \link{style_basename} returns a \link[base]{character} scalar.
}
\description{
Highlight Style for File Base Name
}
\examples{
cat(style_basename('./a/b.R'))
message(style_basename('./a/b.R'))

}
\keyword{internal}
