% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_.R
\name{style_samplesize}
\alias{style_samplesize}
\title{Highlight Style for Sample Size}
\usage{
style_samplesize(x)
}
\arguments{
\item{x}{\link[base]{integer} scalar}
}
\value{
Function \code{\link[=style_samplesize]{style_samplesize()}} returns a \link[base]{character} scalar.
}
\description{
Highlight Style for Sample Size
}
\examples{
30L |> style_samplesize() |> cat()
30L |> style_samplesize() |> message()

}
\keyword{internal}
