% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_.R
\name{text_interaction}
\alias{text_interaction}
\title{Highlight Style for (\link[base]{interaction} of) \link[base]{factor}s}
\usage{
text_interaction(x)
}
\arguments{
\item{x}{\link[stats]{formula} or \link[base]{character} vector}
}
\value{
Function \link{text_interaction} returns a \link[base]{character} scalar.
}
\description{
Highlight Style for (\link[base]{interaction} of) \link[base]{factor}s
}
\examples{
letters[1:3] |> text_interaction() |> cat()
letters[1:3] |> text_interaction() |> message()
text_interaction(~ mrn + dob) |> cat()

}
\keyword{internal}
