% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThomasJeffersonUniv_info.R
\docType{package}
\name{ThomasJeffersonUniv-package}
\alias{ThomasJeffersonUniv-package}
\title{Handy Tools for TJU/TJUH Employees}
\description{
Functions for admin needs of employees of Thomas Jefferson University
and Thomas Jefferson University Hospital, Philadelphia, PA.
}
\keyword{package}
