% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/difftime_ext.R
\name{timeUnits}
\alias{timeUnits}
\title{Additional Time Units}
\usage{
timeUnits()
}
\value{
Function \link{timeUnits} returns a named constant \link[base]{character} \link[base]{vector}.
}
\description{
Additional time units for \link[base]{difftime} object.
}
\details{
To support additional time units \code{'months'} and \code{'years'}.
}
\note{
Functions \link[base]{as.difftime} and \link[base]{units<-.difftime}
only support
\code{'secs'}, \code{'mins'}, \code{'hours'}, \code{'days'}, \code{'weeks'}.
}
\keyword{internal}
