% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phone10_zip5.R
\name{zip5}
\alias{zip5}
\title{5-digit US Zip Code}
\usage{
zip5(x)
}
\arguments{
\item{x}{\link[base]{character} \link[base]{vector}}
}
\value{
Function \link{zip5} returns a \link[base]{character} \link[base]{vector} of \link[base]{nchar}-5.
}
\description{
..
}
\details{
Function \link{zip5} converts all US zip codes to 5-digit.
}
\examples{
zip5(c('14901', '41452-1423'))
}
