% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCount.R
\name{checkCount}
\alias{checkCount}
\title{Positive Counts in a \link[base]{logical} vector}
\usage{
checkCount(x)
}
\arguments{
\item{x}{\link[base]{logical} \link[base]{vector}}
}
\value{
Function \link{checkCount} returns a \link[base]{character} scalar.
}
\description{
Number and percentage of positive counts in a \link[base]{logical} \link[base]{vector}.
}
\examples{
checkCount(as.logical(infert$case))
}
