% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDuplicated.R
\name{checkDuplicated}
\alias{checkDuplicated}
\title{Inspect Duplicated Rows in a \link[base]{data.frame}}
\usage{
checkDuplicated(
  data,
  f,
  dontshow = character(length = 0L),
  file = tempfile(pattern = "checkdup_", fileext = ".xlsx"),
  rule,
  ...
)
}
\arguments{
\item{data}{\link[base]{data.frame}}

\item{f}{\link[stats]{formula},
criteria of duplication, e.g.,
use \code{~ mrn} to identify duplicated \code{mrn}, or
use \code{~ mrn + visitdt} to identify duplicated \code{mrn:visitdt}}

\item{dontshow}{(optional) \link[base]{character} scalar or \link[base]{vector},
variable names to be omitted in output diagnosis \code{file}}

\item{file}{\link[base]{character} scalar,
path of diagnosis file,
print out of substantial duplicates}

\item{rule}{\link[base]{language}, rule of dealing with duplicates}

\item{...}{additional parameters, currently not in use}
}
\value{
The function \code{\link[=checkDuplicated]{checkDuplicated()}} returns a \link[base]{data.frame}.
}
\description{
To inspect duplicated rows in a \link[base]{data.frame}.
}
\examples{
fml = Formaldehyde
fml$id = seq_len(nrow(fml))
(x = fml[c(1:3, 1:6), ])
x$optden[4] = x$optden[1] + 1
x$carb[2] = x$optden[5L] = NA_real_
x$optden[3] = x$carb[6L] = NA_real_
x

checkDuplicated(x, ~ id)
checkDuplicated(x, ~ id, rule = optden == max(optden))

}
\keyword{internal}
