% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_.R
\name{inspect_}
\alias{inspect_}
\title{Data Inspection}
\usage{
inspect_(x, row_dup_rm = TRUE, col_na_rm = TRUE, ptn_Date, ...)
}
\arguments{
\item{x}{\link[base]{data.frame}}

\item{row_dup_rm}{\link[base]{logical} scalar, whether to remove duplicated rows.
Default \code{TRUE}.}

\item{col_na_rm}{\link[base]{logical} scalar, whether to remove all-missing columns.
Default \code{TRUE}.}

\item{ptn_Date}{\link[base]{regex},
regular expression pattern of names of the columns to be converted to \link[base]{Dates}.}

\item{...}{additional parameters, currently not in use}
}
\value{
Function \link{inspect_} returns a \link[base]{data.frame}.
}
\description{
..
}
\details{
\itemize{
\item {convert all eligible \link[base]{integer}, \link[base]{numeric},
\link[base]{factor} and \link[base]{character} columns to \link[base]{logical}.}
}
}
\note{
Be aware of potential name clash, e.g., \code{lavaan::inspect}.
}
