% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ma.R
\name{ma}
\alias{ma}
\title{Moving Average}
\usage{
ma(x, order = 5L)
}
\arguments{
\item{x}{\link[base]{numeric} \link[base]{vector}}

\item{order}{\link[base]{integer} scalar}
}
\description{
..
}
\note{
Function \link{ma} is a simplified version of function \link[forecast]{ma} in package \CRANpkg{forecast}.

Function \link{ma} is much faster than function \link[zoo]{rollmean} in package \CRANpkg{zoo}.
}
\examples{
unclass(ma(1:20, order = 3L))
}
\references{
\url{https://stackoverflow.com/questions/743812/calculating-moving-average}
}
