% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_call.R
\name{Reduce_call}
\alias{Reduce_call}
\alias{do_plus}
\alias{do_average}
\title{Create a \link[base]{call} of Plus}
\usage{
do_plus(x)

do_average(x)
}
\arguments{
\item{x}{\link[base]{character} \link[base]{vector}}
}
\value{
Function \code{\link[=do_plus]{do_plus()}} returns a \link[base]{call}
}
\description{
Create a \link[base]{call} of Plus
}
\examples{
do_plus(letters)

do_average(letters)
}
\keyword{internal}
