% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowSums_.R
\name{row_fullNA}
\alias{row_fullNA}
\alias{rowSums_}
\alias{rowMeans_}
\alias{rowAnys_}
\title{Row and Column Operations Recognizing Full Row/Column Missingness}
\usage{
rowSums_(x)

rowMeans_(x)

rowAnys_(x)
}
\arguments{
\item{x}{\link[base]{array} of two or more dimensions, or a \link[base]{numeric} \link[base]{data.frame}}
}
\value{
Functions \code{\link[=rowSums_]{rowSums_()}}, \code{\link[=rowMeans_]{rowMeans_()}} return a \link[base]{numeric} \link[base]{vector}
}
\description{
To perform row and column operations using \code{na.rm = TRUE},
and only to return a missing value if a full row/column of the input is missing.
}
\details{
Function \code{\link[=rowSums_]{rowSums_()}} performs ..

Function \code{\link[=rowMeans_]{rowMeans_()}} performs ..
}
\note{
Potential name clash with \link[matrixStats]{rowMeans2}
}
\examples{
(x = matrix(c(1, 2, NA, 3, NA, NA), byrow = TRUE, ncol = 2L))
rowSums(x)
rowSums(x, na.rm = TRUE)
rowSums_(x)

}
\seealso{
\link[base]{rowSums} \link[base]{rowMeans}
}
