% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchDF.R
\name{rsplit_}
\alias{rsplit_}
\title{Split \link[base]{data.frame} by Row}
\usage{
rsplit_(x)
}
\arguments{
\item{x}{\link[base]{data.frame}}
}
\value{
Function \link{rsplit_} returns a \link[base]{list} of \link[base]{nrow}-1 \link[base]{data.frame}s.
}
\description{
\link[base]{split.data.frame} into individual rows.
}
\note{
We use \link[base]{split.data.frame} with argument \code{f} being \code{attr(x, which = 'row.names', exact = TRUE)} instead of
\code{seq_len(.row_names_info(x, type = 2L))},
not only because the former is faster, but also \link[base]{.rowNamesDF<-} enforces
that \link[base]{row.names.data.frame} must be unique.
}
\examples{
rsplit_(head(mtcars)) # data.frame with rownames
rsplit_(head(warpbreaks)) # data.frame without rownames
rsplit_(data.frame()) # exception
}
\keyword{internal}
