% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_.R
\name{style_interaction}
\alias{style_interaction}
\title{Highlight Style for (\link[base]{interaction} of) \link[base]{factor}s}
\usage{
style_interaction(x)
}
\arguments{
\item{x}{\link[stats]{formula} or \link[base]{character} vector}
}
\value{
Function \link{style_interaction} returns a \link[base]{character} scalar.
}
\description{
Highlight Style for (\link[base]{interaction} of) \link[base]{factor}s
}
\examples{
cat(style_interaction(letters[1:3]))
message(style_interaction(letters[1:3]))
cat(style_interaction(~ mrn + dob))

}
\keyword{internal}
