% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0PACKAGE.R
\docType{package}
\name{ThomasJeffersonUniv-package}
\alias{ThomasJeffersonUniv}
\alias{ThomasJeffersonUniv-package}
\title{Handy Tools for TJU/TJUH Employees}
\description{
Functions for admin needs of employees of Thomas Jefferson University
and Thomas Jefferson University Hospital, Philadelphia, PA.
}
\author{
\strong{Maintainer}: Tingting Zhan \email{tingtingzhan@gmail.com} (\href{https://orcid.org/0000-0001-9971-4844}{ORCID}) [copyright holder]

}
\keyword{internal}
